requirements.txt
streamlit
opencv-python
matplotlib
pillow
numpy
